Update 2
Updated the PlayerController to include the inventory and weapons list. Player can now pick up objects and those objects are placed in the inventory.
Player can zoom in with right click.
Player has a health bar and can take damage.
Updated the MatchPosRot script to include offsets.
Timescale can be controlled with the - and + buttons, unless we plan to use this as a feature we should remove it before we turn in the final project.
Updated Item to use namespace requirements listed on canvas.

Note: The pickup system can be improved greatly by doing it via Raycast, right now it is done with triggers and gets confused if you are within multiple triggers.