Update 4 PlayerController update
Updated ALL the movement for the player, including WASD movement, jumping, double jump, and jetpack.
Fixed camera jitter when moving/looking around.
The player now slides off of walls and such instead of sticking to them.
Used Will's level 1 and removed all light except for a point light at the center of the room on the ceiling.
The DevLevel now has a green door that you can touch to teleport to Level 1.