Shop and Music
Added a Folder in Imported Assets named; r_MiscAssets to  temporarily hold the assets I am bringing in, of which I added the Song as well as the Image for the Shop.

Started to mess around with the Main Menu as well as the Shop again. Furthermore I started editing the "Settings" script which I made previously to save different saved settings (Mainly Sens) inside of the Main Menu.