Music
Added in a Temp Assets File in ImportedAssets for me to start putting things in.

Attached the song to the Player Controller in Main Menu, upon trying to do this with the Prefab I had gotten an Error Message, dunno if there was something else I was supposed to do.