Main Menu UI
Added in the Sensitivity Slider, if the Programmers could set that up to actually affect the PlayerController for me I'd be over the moon.

Besides that I also added the Options menu, which basically just disabled and enables whatever menu is open ie; the "Start" menu and the "Options" menu.

Also removed a script that I made a bit ago called Settings, which I had thought of using to implement a Save/Load feature that was demonstrated to me and a few of my peers in Cameron's Class, however, I obviously wouldn't have enough time to implement that.