Shop Menu
Started work on the Menu inside of Shop Menu, I see we already have something set up alongside the Scenes ShopMenu2 & 3 or something.

Added a starting screen that would load first upon completing the mission telling you the Credits earned, then as well as a button that sends you to the actual Shop.

Among the shop, the only real thing I have so far is just a couple of buttons that would send you to the next level and one that would send you back to the main menu.

Go ahead and tell me if anything is overly ambitious for the little time I have to develop this. I realize that if I had been more available we could have set it up in such a way that we can go through level selection which might be too much for the moment.